+++ python/lsst/ts/mtdomegui/__init__.py
@@ -34,6 +34,7 @@ from .constants import *
 from .control_panel import *
 from .main_window import *
 from .model import *
+from .reporter import *
 from .signals import *
 from .status import *
 from .tab import *
